PyPySide: circumvent an lldebug regression
When building PyPy with lldebug or lldebug0 selected a
a build option, there is a regression in the signature
module that crashes every build

This crash is a confirmed PyPy bug with global
dictionaries that occurs sometime and is hard to fix
on the PyPy side.

The workaround is to skip signature initialization in
such a PyPy debug build.

Task-number: PYSIDE-535
Change-Id: I9977e5ff82759bfd882d55785528dd276012e047
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>