__feature__: handle properties with function overloads
This is the implementation, see the conclusion of the issue.

[ChangeLog][PySide6] When a property would override an
existing function with multiple arity or parameters,
append an underscore to the property name.

REMARK: The current implementation is very correct. It uses
introspection via the signature module.
But that adds a constant overhead to the true_property
feature. Actually, there are only 2 known cases where this
overlap happens. It might be considered to simplify things
by checking the string names of these two functions?

Fixes: PYSIDE-1670
Pick-to: 6.2
Change-Id: I14927995698726957ba8c515dddf5e37c21910ce
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>