CMake: Cleanup documentation building code
Move conditions on whether documentation should be built inside the
doc project, rather than outside.

Look for qdoc and qhelpgenerator both in PATH and via
find_package(Qt6Tools) calls.

Add sanity checks that the various tools used in the
add_custom_command calls are actually available.
Show a warning or error when they are not available,
depending on combination of whether it's a standalone doc
build or part of the overall shiboken/python build.

Set DOC_OUTPUT_FORMAT to html by default if it's not specified.

Clean up some of the warning messages.

Make sure to check the result of running execute_process on the
snippets tool.

Pick-to: 6.2
Task-number: PYSIDE-1718
Change-Id: I2969d7a40961881ad0f91d7142b29a7f4130f3b7
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>