Revert "setup.py: Move all build dirs into a common 'build' subdirectory"
This reverts commit 234349d124ccfa399921e2b9fc09addcff0b0a94.

It fixes the following warning
 warning: PysideInstallLib: 'build/lib.macosx-10.14-x86_64-3.7'
          does not exist -- no Python modules to install

The build_lib directory is only assigned during the run() method of
the build command, which means the install command didn't pick up
anything to install, because it expects the build_lib to be set during
finalize_options of the build command, and it wasn't, it had the
default value instead.

Installation accidentally worked for the bdist_wheel command because
the build command is run to completion before the install command
is prepared.

Pick-to: 6.2
Change-Id: I512a958db30858344f0d873a23bff0b925d7618e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>