shiboken6: Implement enum search by value in code model
Replace (currently unused) AbstractMetaBuilder::searchForEnumScope()
by _ScopeModelItem::findEnumByValue() returning an enum and the fully
qualified name.

This function is intended for resolving enum values used as default
values for functions taking int where the underlying enum is not
known.

As opposed to old code, all parts of a (partially) qualified name are
checked for a match to ensure no mismatches of equally named values.

Another advantage is that also enum values that are not in the type
system are found.

The function returns the fully qualified name (also including
the enum name for non-class type enums).

Task-number: PYSIDE-1691
Pick-to: 6.2
Change-Id: I89ebfdf8435470c626cfdee4fc0d0738cc3fc195
Reviewed-by: Christian Tismer <tismer@stackless.com>