__feature__: Fix snake_case handling on user defined classes
The snake case feature filters candidate methods and turns
them into snake case. This works fine for built-in classes.
The assumption is that all methods come from the tp_methods
list.

This assumption is not correct when applied to user defined
classes. The methods have no static source in this case.
To distinguish here, we inspect the tp_methods list. If it
is empty, we assume a user defined class and do nothing.

A forgotten initialization in feature.py was added, too.

As a note: RHEL has such an old Python version that does
not have MethodDescriptorType in the types module.

[ChangeLog][PySide6] snake_case handling now does explicitly
not touch user defined classes.

Fixes: PYSIDE-1702
Change-Id: Idfa16cdc50cb7234c1d2f473dfae3a568887547e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 42695262f0ab4c73377b9d638dd28636ab1a3668)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>