PyPySide: Replace type patching by heaptype, 1(5)
PyPy does not allow patching of non-heaptypes. We did that
trick for simplicity in the signature module where we created
a `__signature__` attribute for types.

This patch removes the type modification and uses our
heaptypes, only. For other types, more work is needed.

It also fixes the metatype for enums which was wrong.

This is the first replacement of

    o PyType_Type
    * PyMethodDescr_Type
    * PyCFunction_Type
    * PyStaticMethod_Type
    * PyWrapperDescr_Type

The patches will be replaced by heaptypes if possible,
or we might use signatures in a shadow dict, instead.

[ChangeLog][shiboken6] Remove type patching of PyType_Type
and fix the meta type of enum.

Task-number: PYSIDE-535
Change-Id: I221834661709c71ee9ed17b3d6cc293b87447ab7
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>