mypy-support: Improve handling of slots, enum and PathLike
As a preparation of the move to Python enums, the types
of enums are now reported as Python enum. This avoids certain
errors in the .pyi files. The correct implementation will
follow.

PathLike is now correctly imported via os if necessary.
slot() has corrected signature and reports a call method.

[ChangeLog][shiboken6] The .pyi file generation was improved to
have correct slot(), PathLike and better enum handling.

Change-Id: I547de6a1203302f73826f27ee633e9de810f687b
Fixes: PYSIDE-1520
Task-number: PYSIDE-1675
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>