shiboken6: Handle PySets as function parameters
Python sets are iterable but not of sequence type.

While the existing converter code from the templates uses iterators,
the built-in check functions
convertibleSequenceTypes()/checkSequenceTypes() assume a PySequence
and would fail for PySets.

Add new check functions convertibleIterableTypes()/checkIterableTypes()
using iterators and use them for PySet.

Add a test and a test for lists as a drive-by.

[ChangeLog][PySide6] sets are now supported for functions taking a QSet.

Pick-to: 6.2
Task-number: PYSIDE-174
Task-number: PYSIDE-1666
Change-Id: I883869162e7dfa9cd0e1669f941fb7864f0cf825
Reviewed-by: Christian Tismer <tismer@stackless.com>