PyPySide: Fix BindingManager::getOverride and others
PyPy has a major difference when it comes to distinguish
PyCFunction and PyFunction/PyMethod. PyCFunction is not
differentiated in PyPy and its handling collapses with
PyMethod handling.

A simple way to emulate the difference is by a redefinition
that uses a check for __module__, that does not exist for
builtins.

Observation: There are quite a number of false callbacks in
Qt that create wrong notifications. Before this error was fixed,
these errors were reported by the error handler. When handled
correctly, these spurious notifications are ignored, again.

The effect of this change is very good:
76 errors left instead of 99.

[ChangeLog][shiboken6] The override handling for PyPy was fixed
because PyPy has a different understanding of builtin functions.

Task-number: PYSIDE-535
Change-Id: I7d7419135a50df9a540e953b75890ca2ec28993c
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>