setup: Make ninja output the commands it runs in a verbose build
Verbose build output worked when using the Makefiles generator because
we set CMAKE_VERBOSE_MAKEFILE to ON, but that does not affect ninja.

For ninja we need to explicitly pass -v on the command line.

The verbose ninja output is useful when debugging build issues in the
CI (which uses verbose build).

Change-Id: Ib1532db0225744184d89bf796c4b3a6a40d718ca
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 9594789fcc58cb250900c935e1adb4b2ce9ed8b6)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>