setup: Make ninja output the commands it runs in a verbose build
Verbose build output worked when using the Makefiles generator because
we set CMAKE_VERBOSE_MAKEFILE to ON, but that does not affect ninja.

For ninja we need to explicitly pass -v on the command line.

The verbose ninja output is useful when debugging build issues in the
CI (which uses verbose build).

Pick-to: 6.2
Change-Id: Ib1532db0225744184d89bf796c4b3a6a40d718ca
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>