shiboken6: Refactor Python operator handling
The mapping of the function name to the Python operators was duplicated
in ShibokenGenerator and QtDocGenerator.

Move it to the Generator base class and use it in the QtDocGenerator
as well. Add the underscores.

Remove the functions retrieving the Python rich comparison operator
code from ShibokenGenerator and add a comparison operator type enum to
AbstractMetaFunction instead. Use that to retrieve the Python rich
comparison operator code.

Task-number: PYSIDE-1711
Change-Id: Ib73412b819c60c3af22bc72c6bd1cfaa7f25904a
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>