Reland: Move all build dirs into a common 'build' subdirectory
When building the project using setup.py, it would pollute
the source directory with a lot of virtual environment build
directories.

Move all of those into a common 'build' subdirectory.
This eases cleanup of all build directories because they
can all be removed with a single rm command.
It also places all the various stages of build -> install -> package
folders into a common subdirectory so it's easier to find and navigate
between them.

If no virtualenv is detected, change the venv prefix from the previous
value of 'pyside' to 'qfp' and include the build classifiers.
If a virtualenv is detected, we consider it is distinct enough and
don't include the build classifiers, apart from a few chars to denote
a debug python, debug Qt or limited api build.

Example of the new build directory structure when a virtualenv is
detected

 build/{venv_name}/build/shiboken6
 build/{venv_name}/build/pyside6
 build/{venv_name}/install/bin/shiboken6
 build/{venv_name}/package

Example of the new build directory structure when a virtualenv is
NOT detected

 build/qfp-py3.9-qt6.2.0-64bit-release/build/shiboken6
 build/qfp-py3.9-qt6.2.0-64bit-release/build/pyside6
 build/qfp-py3.9-qt6.2.0-64bit-release/install/bin/shiboken6
 build/qfp-py3.9-qt6.2.0-64bit-release/package

Move the code that always removed the ./build directory on each
setup.py invocation into prepare_packages() instead.
This way it only removes the files from the 'package' subfolder which
is the common packaging location between all the sub-projects.
This removal is needed to ensure shiboken files don't end up packaged
in the PySide6 wheel.

This relands commit 234349d124ccfa399921e2b9fc09addcff0b0a94

This reverts commit 0c6eb7cd232fff9d81a8d5bc9a7fd71d9b8c67f5

[ChangeLog][setup.py] Build directories are now created inside the
root ./build directory, rather than directly under the root of the
project.

Pick-to: 6.2
Change-Id: I6d511ae77cb66c2c5a872d6b85ff33e1831b803e
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>