shiboken6: Simplify generating of error returns
shiboken6 needs to generate code with different error returns, for
example PyObject *nullptr in normal wrappers and int -1 in _init()
functions.

For this, CppGenerator had a static string member that could be
overridden by instantiating a RAI class ErrorCode. Replace this by a
simple enumeration representing the error codes and pass it to the
functions where needed.

Change-Id: I785ef79ecc25303a87f12959f91f83178c64da1d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>