signature: improve error handling for embedded applications
Entering something like

"""
mainWindow.setPointer(None)
"""

crashes in an old version of scriptableapplication, which
shows an omission in the signature interface. The error
shows up whenever a builtin module cannot be imported.

The error does not show up in PySide 6, because the
module is declared as a builtin via `PyImport_AppendInittab`.

* add registration of AppLib as a builtin (5.15)

* insert builtin modules per default into the mapping module

* simple recovery if a module cannot be found in sys.modules

[ChangeLog][shiboken6] Error handling was improved for embedded
applications and builtin modules are trusted as valid modules.

Change-Id: I722212a52a5e3aae924f0b965578485ecaf185a9
Fixes: PYSIDE-1710
Pick-to: 6.2 5.15
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>