setup.py: Fix no log messages being shown during setup call
The default verbosity of a log object is WARN. distutils then set it
to INFO when initializing the Distribution object. This would not
affect the log object copy that setuptools exposes (and we now use).
This caused the usual INFO messages not to be shown anymore.
Explicitly set the setuptools log object verbosity to INFO unless the
quiet option was given.

Amends 95a5bb9dd3b5d3fa86f2ed0868e2b821256a6028

Pick-to: 6.2
Change-Id: I793dc92582007895fa23d43baabe5b97c146552e
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>