snippets_translate: Streamline writing files
The code created all example directories first and extracted the
snippets to .cpp files which were later moved to .py files to
overwrite the snippets in the repository (back then).

Write to the Python file directly and create the directory only when
needed. This removes a lot of empty directories.

Pick-to: 6.2
Task-number: PYSIDE-1721
Change-Id: I38e695db4aae386be18cd3b85708f684bfa228d5
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>