Documentation: Fix snippets resolution
Change snippets_translate to replicate the complete Qt tree
including module name under the codesnippet directory.

Resolving the snippet is then a matter of replacing
the Qt source path by the PySide 6 snippet path in the fallback
paths contained in WebXML and checking for the converted .py
files.

This basically reduces the snippet warnings to 0.

Pick-to: 6.2
Task-number: PYSIDE-1721
Change-Id: I344004ddd1580f96ebf0c132923388ad08852374
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>