Fix build when Vulkan is not installed
The Qt headers define some dummy types for Vulkan types that take
effect when Vulkan is not installed. This causes
QQuickRenderTarget::fromVulkanImage() and some other functions to be
declared with no linkage, causing a link error in PySide. Reject the
function (which happens in the normal case since VKImage is not known
to shiboken). Suppress the warnings in the normal case.

Pick-to: 6.2
Fixes: PYSIDE-1726
Change-Id: I915b477280bac5200c8206a7f9044b78c80e2936
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>