mypy-support: Improve the handling of `Signal`
After more correct `Slot` handling, `Signal` now gets a __get__
method that always was there, but did not show up in .pyi files.

By this `__get__` function and its return-type `SignalInstance`,
the correct connection to `emit`, `connect` and `disconnect`
is automatically there.

[ChangeLog][PySide6] The MyPy display of `Signal` was corrected
so that the relationship of `Signal` and `SignalInstance` is reflected.

Task-number: PYSIDE-1675
Change-Id: I58e66008418147f37e0b96067d62677b6b0b902c
Fixes: PYSIDE-1318
Pick-to: 6.2
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>