Signature: fix the __doc__ attribute of classes
The signature module had been changed to no longer default
the __doc__ attribute of classes to the __init__ signature.
This has the side effect of crashing "help(QtCore)".

Fixed by correct defaults in C++ (AttributeError) and by
setting a "None" default in the Python handler.

The make_helptest function defaults again correctly to the signature:

    >>> errorhandler.make_helptext(QtWidgets.QApplication)
    'QApplication(self) -> None\nQApplication(self, arg__1: Sequence[str]) -> None'

Change-Id: I140f2b940f98eb126541b18b0feb312c7c4e9728
Fixes: PYSIDE-1727
Pick-to: 6.2 5.15
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>