example_gallery: Handle dummy projects in the doc directory
Make it possible to have a dummy pyproject file with entries
pointing to the parent directory in the doc directory as not
to clash with a CMakeLists.txt in the examples directory.

Separate the code paths for the pyproject_file and the actual
examples directory for this purpose.

Change-Id: I0a1e583bff8cbb8243ba7526b8b2908f6a3e6e9c
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 8cbfe9aa6f46483968b8273aa45deaad12bffbfd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>