PySide6: build multiple inheritance pass-through for Signals
Signals always had support for calling a method that is sitting
at the same place as a signal with the same name. This is done
by letting the signal act as a proxy for the hidden method.

Lately, the equality check of signal instances was corrected
(PYSIDE-1431) which had the side effect that now some method
lookups failed in multiple inheritance.

This patch extends the search for homonymous methods to look
not only in the signal attribute, but extends the lookup
along the whole tp_mro.

The related tests were slightly re-arranged to keep them together.

[ChangeLog][PySide6] Signal instances allow pass-through of methods
with the same name, explicitly and with multiple inheritance.

Change-Id: I92828855ada7aa8d98517918008b93fc79629d8e
Fixes: PYSIDE-1730
Task-number: PYSIDE-1431
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 85799120d09e251a761699ff71547b7dc80a741b)
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>