setup.py: Prevent creation of byte-compiled files
Byte-compiled files (.pyc or .pyo) are generated by
default when a module gets imported.
The "setup install" command has additionally the effect
of pre-compiling all Python files for the installation.

This is much less relevant than when this behavior was
implemented (was in Python 1.4, already in 1996).

We don't want this behavior that clutters example directories
and does not make much sense when the Limited API is used.
The compiled files will still be created on import.

But be aware of possible side effects because files may
be written on a read-only installation.

[ChangeLog][PySide6] The byte-compiling of example files is
now suppressed to save space and clutter.

Task-number: PYSIDE-1746
Change-Id: I811431030517c251f32bcadc4c98fb646b68eafa
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>