build_scripts: Avoid a racing condition in install
On macOS and a custom Python built, the "install" option
crashes with an import error.

This is a timing problem, because an import is triggered
when it is already being processed. The problem is hidden
by the use of virtualenv.

[ChangeLog][PySide6] The move from distutils to
setuptools caused a racing condition with certain
Python builds.

Change-Id: I4dcda2b083234e88c272a1794388f6d0e7b8f0e5
Fixes: PYSIDE-1760
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>
(cherry picked from commit 981ffb345928cf740a5b751aaef43a266d077d13)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>