signal: Fix equality test for inherited signals, update
Signal initialization was improved by walking the MRO
to find all signals. But this had the side effect that
signals could be overwritten.

This implementation resolves signals only once per name,
so that only the most recent signal places an instance
into the instance dict.

Note: this does _not_ solve the equality problem of

    super().signal == super().signal  # fails

because the super call triggers a new signalInstance in the
parent. This has to be addressed in a bigger rewrite if desired.

[ChangeLog][PySide6] Signal initialization now respects
the order of subclasses in the mro().

Change-Id: I7481576798ca272ebebf8e9a5ba875cfc94e8361
Fixes: PYSIDE-1751
Pick-to: 6.2
Task-number: PYSIDE-1431
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>