Safe distinction of Nuitka compiled methods
Adds an extra check to see if __code__ is present.

As mentioned in PYSIDE-1755, Mocks are callable objects without __code__
attribute, unlike Python Method or Functions. However, a Mock also has
im_func and im_self attributes. We made the assumption __code__
would be present if im_func and im_self are present, and this makes it
fall under the category of a compiled method.

This patch makes an extra check to see if __code__ is present. If it is
not, then the Slot (here Mock) is considered as a callable method.

Task-number: PYSIDE-1755
Change-Id: If7e8f52dfb2409cd856eec0d0b41891d751d8a69
(cherry picked from commit 56f66f128566bd08f027fee46bb42a6c4f57a26e)
Reviewed-by: Christian Tismer <tismer@stackless.com>