shiboken6/Type system parser: Introduce a helper for the code snippets
Introduce a helper returning a parser state depending on the element
stack and building on that, a helper which returns the code snippet to
append to. This is used for character handling and <insert-template>.

It also allows for removing StackElement::InjectCodeInFunction,
which was a fake element for remembering that the code injection
targets a function.

Also, add all type entry code snippets directly to the type entry
instead of temporarily storing them in StackElementContext::codeSnips.
Use the latter for conversions only and rename it accordingly.

Task-number: PYSIDE-1766
Change-Id: I291ec2db58687fdf63c124ddde26165f8eb09856
Reviewed-by: Christian Tismer <tismer@stackless.com>
(cherry picked from commit b4ffbed19625bbf468c660a6dcc35a1101b8791b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>