shiboken6: Fix handling of value types with protected constructors
Value types with some protected constructors (like
QOperatingSystemVersionBase in 6.3) caused compile errors on Windows
where the protected hack is disabled since non-accessible constructors
were used.

The check for isValueTypeWithCopyConstructorOnly() needs to be fixed
to exclude protected constructors in case AvoidProtectedHack is set.

Similarly, the visibility of the minimal default constructor needs to
be checked. Add the AvoidProtectedHack setting to ApiExtractorResult
for this purpose since it influences the API in a way.

Pick-to: 6.2
Change-Id: Ifeab320b9391aa21c5b1de4d21d2d8276fe44d3f
Reviewed-by: Christian Tismer <tismer@stackless.com>