PyPySide: Rename interface functions and classes to simplify debugging
The names of certain interface functions are not always
following a simple scheme. Especially it is not easy
to see immediately if we are dealing with a method
of SbkObjectType or SbkObject

Do a few renamings to simplify debugging and make the code
easier to understand. When a function is used in a type spec
and there is no other important reason, it should be named
like

    {Py_<tpname>: reinterpret_cast<void *>(<TypeName>_<tpname>)},

Rename also all type functions ending on "TypeF()" to end
in "_TypeF()". This is not always the case.

Examples:
    SbkObjectTpNew -> SbkObject_tp_new
    SbkObjecttypeTpNew -> SbkObjectType_tp_new

    PyClassPropertyTypeF -> PyClassProperty_TypeF

Task-number: PYSIDE-535
Change-Id: Icbd118852f2ee732b55d944ed57c7a8ef7d26139
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>