PyPySide: Fix locker handling which enables Mandelbrot
Not only the QPainter, but also the QMutexLocker
were not correctly modelled after PEP 343 in the
examples. Since that is now fixed, we consider the PyPy
project not as ready, but good enough to publish it.

It also turned out that people have the expectation
to use QSignalBlocker in the "as" form:

    with QSignalBlocker(self.double_spin_box) as blocker:
        self.double_spin_box.setValue(2.5)

https://stackoverflow.com/questions/60384734/how-to-use-qsignalblocker-in-python

But that blocker would be None.

As a side effect, QMutexLocker, QReadLocker, QWriteLocker
and QSignalBlocker were augmented with a default
__enter__ implementation that returns the locker instance.

[ChangeLog][PySide6] The Mandelbrot example needed context managers
for QPainter and QMutexLocker to work in PyPy.

Task-number: PYSIDE-535
Change-Id: I2a62ca645a4fddcafbf11869f14a538141f32c39
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit bc11e3c074bdba1ae0fcaa0c10e3b32f781be9fd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>