libpyside: Factor out helpers for class decorators with parameters
Class decorators with parameters (as opposed to parameterless
decorators which are simple functions) requires Python types
with call operators to be registered.

Add Decorator C struct with a pointer to a class derived from
DecoratorPrivate that has virtual functions for the
tp_init()/tp_call() functions.

Provide templates that help to create the slot arrays for the type
registration.

Provide predefined types for decorators with string/type parameters.

Task-number: PYSIDE-1709
Change-Id: I2946b4505c01b49532af534adcedff7a51634929
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Shyamnath Premnadh <Shyamnath.Premnadh@qt.io>