Convert QVariant to Enum types
Before this patch, the conversion from QVariant->enum in Shiboken was
same as the conversion from QVariant->int. So, the user will have to
explicitly convert the returned integer to the original enum type.

With this patch, an enum type is considered as an Object type rather
than an int type. This enables the enum type to be considered as itself.

Added tests for all conversions of QVariant types.

Pick-to: 6.2
Fixes: PYSIDE-1798
Change-Id: I18acc08e1f59ac7df844444d0ad6950e1d70ab21
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>