PySide: Leave the underscore variable alone in QApplication
QApplication has special treatment through the extra rules
of only one QApplication instance may exist. To ensure this,
different approaches were used.

The first attempt was using refcounts and made sure that
"del qApp" really destroys the QApplication instance.
This did not work in interactive mode because the automatic
underscore variable creates cycles.

As a simple workaround, we assigned None to underscore
when a QApplication was created. When moving to the new
concept of an explicit shutdown() call, this assignment
was forgotten to remove.

[ChangeLog][PySide6] A left-over replacement of the interactive
underscore variable when QApplication is created was removed.

Change-Id: I78f3e3714531d599c4b5b76f956b1d631014ad94
Fixes: PYSIDE-1758
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 0a12054f64f94807f711d35d26091b64f1589b3b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>