PyPySide: Remove a disambiguation patch no longer needed
CPython always had a differentiation between PyMethod and
PyCFunction which PyPy did not have and regarded as quite
artificial. When NumPy suddenly started to grow warnings about
this in a recent version, the PyPy team quickly added such
a distinction, and we get rid of a patch. :)

This patch was introduced on 2021-11-09:
"PyPySide: Fix BindingManager::getOverride and others"

[ChangeLog][PySide6] After PyPy lerned to distinguish between
PyMethod and PyCFunction, we could remove a patch.

Task-number: PYSIDE-535
Change-Id: I441ae7756331b5deccf67e261a7919e4d3e1e1f9
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>