__feature__: Fix true_property overriding
It was assumed that an override of a property would not
make any sense, and the feature was disabled.

An example showed that it _does_ make sense, and this
works without any effort.

This does not fix other problems which are still there,
like correct handling of renamed implicit function calls.
This will be addressed in another change.

[ChangeLog][PySide6] true_property overriding was enabled.

Change-Id: Ic22f05c6c999e7f97c47161d95e785c952168bb6
Fixes: PYSIDE-1765
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>