__feature__: Do the renaming also in the hidden methods
It has taken quite a while until it was understood
why mistyped (un-renamed) methods are still found
when using `snake_case`. It is a very old patch
that searches hidden objects via the QMetaObject
when normal attribute retrieval fails.

This implementation is used all the time, but it is
easily overlooked because most generated wrapper
objects use this implementation from QObjectWrapper
by inheritance, and is not found in the source code.

Renaming these hidden methods the same way as the
normal methods are renamed solves this problem.

As a side effect, the solution to PYSIDE-1765 becomes
more complete, too, since these hidden attributes
were really missing.

[ChangeLog][PySide6] The snake_case feature is now
more complete since it renames hidden methods, too.

Change-Id: I8614314e417a40549ef09f3105a03f22cb575bab
Fixes: PYSIDE-1753
Task-number: PYSIDE-1765
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>