__feature__: rework constructor handling of both features
This is a quite complicated patch that tries to do the
right thing in all cases with snake_case and true_property.

Special is that a property can be a true property
that is really turned into a property object, or it can be
a pseudo-property which is not a property but should be
treated as such.

Exhaustive tests for all cases are added.

[ChangeLog][PySide6] Properties are now usable in constructors
of classes in any combination of snake_case and true_property.

Task-number: PYSIDE-1019
Fixes: PYSIDE-1705
Pick-to: 6.2
Change-Id: I1c6b6a24e0bb2e906ecbed225953ac782fed48f7
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>