PyPySide: Fix a regression with thread initialization
The mandelbrot.py example worked fine on macOS, but failed
to run on Windows and Linux.

This did not show up on my local Windows virtual machine
after building. But cloning the machine and removing the
developer partition revealed the same issue on the clone.

It turned out that PyPy needed a PyEval_InitThreads call.
This call is already deprecated and a no-op on Python 3.9.
Adding this call fixed the problem.

We need to contact the PyPy group because they have a
PyPython 3.9 beta, which is dependent from a deprecated function.

[ChangeLog][PySide6] Threading stability was much improved
by a call to PyEval_InitThreads (deprecated in Python 3.9).

Task-number: PYSIDE-535
Change-Id: Ibf6d92bf0b21542be8929e12a6e02389c036ec79
Pick-to: 6.2
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>