build: script to create wheels
This approach intends to avoid modifying the current
structure we have in build_scripts, and can replace
the call:

    python setup.py bdist_wheel

mainly encouraged by PEP517, and the need of having
incremental wheels, to replace the current single PySide6 one.

The current configuration allows to create two new wheels:
  PySide6_Essentials, and
  PySide6_Addons
that contain all the essential and addons Qt modules defined
by the Qt Installer tool, with some modifications due to the
dependencies of certain tools. Check the README files for more info.
The known PySide6 wheel is also generated, but it's empty in favor
of using the previous two wheels as requirements, installing them
automatically, to avoid modifying the usage of 'pip install pyside6'

The strategy is based on the current logic behing 'prepare_packages'
that we have been using. Once the modules are built, instead of
removing those directories currently in 'build/your_env/package',
we rename them.
Inside this new directory, one can have the 'shiboken6',
'shiboken6_generator', and 'PySide6' directories, with eveything
already packed with the required wheel structure.

The main difference is that instead of using the content of
PySide6 to build one build, we select some files with the MANIFEST.in
to create another wheel.

The wheel tag drops the old assumption of needing:

  cp36.cp37.cp38.cp39.cp310-abi3

and only uses:

  cp36-abi3

Additionally, for Linux, we follow PEP600 to use the GLIBC version
in the wheel name instead of manylinux1, manylinux2010, etc...
For the current CI configuration, we know we are using 2.28, which
is the minimum supported version for Qt6, so the wheel will look like:

  PySide6-6.3.0-cp36-abi3-manylinux_2_28_x86_64.whl

The coin scripts were configured as well, to add the call of the
new create_wheels.py script, and test them via wheel_tester.py

Note: This script is not intended to be used as a general purpose
wheel creation tool, and it's purely focused on the current Qt CI.
There are many ad-hoc configurations used in different functions,
like the structure of a Qt installation, the usage of 'a' on the
environment for limited-api, etc.

Task-number: PYSIDE-1115
Fixes: PYSIDE-692
Change-Id: Ic12e428b8b9b64bbe2facb1c520595ccd2384497
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>