generate_pyi: Fix an import timing problem on Windows
The generate_pyi output is tested by importing it.
On Windows and in a venv, this creates an import cycle.

This error could be attributed to PyPy as well
because the bug was identified this way, but the
reason for the trouble is not PyPy.

The same effect happens with Windows, Python version > 3.6,
venv or virtualenv and the setting "set QTEST_ENVIRONMENT=ci".

Re-ordering imports solved the problem.

Task-number: PYSIDE-535
Change-Id: I3dfca640afa5e2afd36b9fff3315e02447408dbd
Reviewed-by: Christian Tismer <tismer@stackless.com>
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
(cherry picked from commit 0a9c75485ddab77bc1d01d731d5559cec3539d5c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>