testing: fix a bug that prevents commenting entries
When trying to fix PyPy by blacklisting, it was not possible
to blacklist one bug. After a while it was clear: The entry in
question was already present in another, PyPy unrelated position.

The PyPy entries are at the end of blacklist.txt, and so the old
entry was replaced. This is a drawback of the system that could be
discussed.

The problem: As a workaround and to keep things reproducible,
I added a comment to the old entry and repeated the entry in the
PyPy section with a comment, too.

The final problem after long analysis was that comments were no
longer possible since 2021-01-05, by a wrong cosmetic change.
(2de3a763fa18d5c30c2cff25057f1b81ceeed231)

Change-Id: Ibb34a98f010294f03fae7e478c40e5c8742da689
Pick-to: 6.2
Task-number: PYSIDE-535
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>