testing: Allow for repetition of blacklist sections
The testing framework was written with the rule that
sections may not repeated - only the last entry would
persist.

When adding a whole row of PyPy entries, it turned out
to be a drawback. It is nicer if sections can be repeated
without interaction.

[ChangeLog][shiboken6] In blacklist.txt now sections can
be repeated. They will be combined as one section.

Task-number: PYSIDE-535
Change-Id: I5e2ac435c7508b542306b470eca50411ee01606b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>
Reviewed-by: Cristian Maureira-Fredes <cristian.maureira-fredes@qt.io>